// cypress/support/types.ts
export enum InsightType {
  Article = "Article",
  Webinar = "Webinar",
  Whitepaper = "Whitepaper",
  MarketUpdate = "MarketUpdate",
  QuarterlyReview = "QuarterlyReview",
  ManagementView = "ManagementView",
}

export type Insight = {
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
}

export type Insights = {
  type: InsightType
  insights: Insight[]
}

export type InsightsWithPagination = {
  page: number
  totalCount: number
  data: Insight[]
}

export type InsightContent = {
  title: string
  description: string
}

export type Article = {
  details: string
  contents: InsightContent[]
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  disclaimer?: string
}

export type Guest = {
  id: string
  name: string
  picture: string
  title: string
  type: string
}

export type Webinar = {
  bannerImage: string
  cardImage: string
  description: string
  estimatedDuration: string
  id: string
  publishDate: string
  title: string
  contents: InsightContent[]
  video: string
  guests: Guest[]
  insightType: InsightType
  disclaimer?: string
}

export type MarketUpdate = {
  details?: string
  contents: InsightContent[]
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  disclaimer?: string
  downloadLink?: string
}

export type MonthlyMarketUpdate = {
  details: string
  contents: InsightContent[]
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  disclaimer?: string
  downloadLink?: string
}

export type ManagementView = {
  details: string
  contents: InsightContent[]
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  video: string
  disclaimer?: string
}
// ...
export type QuarterlyReview = {
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  downloadLink: string
  contents: InsightContent[]
  disclaimer?: string
}

export type Whitepaper = {
  insightType: InsightType
  id: string
  title: string
  bannerImage: string
  cardImage: string
  description: string
  publishDate: string
  estimatedDuration: string
  downloadLink: string
  contents: InsightContent[]
  disclaimer?: string
  video?: string
}